Start simplifying consumer
