+++ src/main/java/net/parttimepolymath/spring/springkafka/configuration/Config.java
@@ -1,7 +1,5 @@
 package net.parttimepolymath.spring.springkafka.configuration;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -13,17 +11,9 @@ import org.springframework.context.annotation.Configuration;
  */
 @Configuration
 public class Config {
-    @Value("${default.count}")
-    private long defaultCount;
-
 
     @Bean
     public RuntimeConfig runtimeConfig() {
         return new RuntimeConfig();
-    }
-
-    @Bean
-    public CLIParser cliParser(final @Autowired RuntimeConfig config) {
-        return new CLIParser(config, defaultCount);
     }
 }
